+++ CHANGES.md
@@ -77,7 +77,7 @@ Version PBFT
  -  (Libplanet.Net) Added `Message.Id` property.  [[#PBFT]]
  -  (Libplanet.Net) Added `Gossip` class.  [[#PBFT]]
  -  (Libplanet.Net) Added `Proposal` class.  [[#PBFT]]
- -  (Libplaent.Net) Added `ProposalMetaData` class.  [[#PBFT]]
+ -  (Libplaent.Net) Added `ProposalMetadata` class.  [[#PBFT]]
 
 ### Behavioral changes